Import era from sod classes
